feat: teardown preview on pull request closed
